adding readme and example env (#137)
This pull request includes significant updates to the `.env.example`
file and a comprehensive overhaul of the `README.md` file to better
reflect the purpose and usage of the PyPSA-Earth Visualization
Dashboard.

### Environment Configuration:
*
[`.env.example`](diffhunk://#diff-a3046da0d15a27e89f2afe639b25748a7ad4d9290af3e7b1b6c1a5533c8f0a8cR1-R6):
Added environment variables for PostgreSQL configuration.

### Documentation Enhancement:
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R62):
Updated the introduction to describe the PyPSA-Earth Visualization
Dashboard, its purpose, and data sources.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R62):
Added detailed sections on the main pages of the dashboard, including
Home Page, Now Page, and 2050 Page, with visual previews.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R62):
Included a Contributors section with links to contributor profiles.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R62):
Provided instructions for building and deploying the application.
*
[`README.md`](diffhunk://#diff-b335630551682c19a781afebcf4d07bf978fb1f8ac04c6bf87428ed5106870f5L1-R62):
Added links for more information about PyPSA-Earth and related
resources.

### Home and Learn More 
@ElectricMountains would you like to update the text for home and learn
more page ?