Charts Information Tooltip (#150)
This pull request includes several changes to the `app/landingglobe` and
`components/Charts` directories to improve the user interface and add
new tooltip information. The most important changes include the removal
of unnecessary imports, the addition of new tooltip components, and
updates to the drawer components to enhance the display of information.

utilities/TooltipInfo/ExplainerText/GenerationMix.ts

files now cotains text which will be displayed in tooltip