Globe section -  Electricity Price in system cost drawer,Globe Legend,investments on hover (#153)
This pull request includes significant changes to the `LandingGlobe`
component and its associated popups. The major updates include the
introduction of a new `GlobeLegend` component, removal of the
`ElectricityPriceComponent`, and refactoring of data fetching and
display logic.

### New Component Addition:
*
[`app/landingglobe/popups/GlobeLegend.tsx`](diffhunk://#diff-b93cdc3c718059683ea43de2cd89f083d8a845884828ff04c57a1c908c8245e2R1-R29):
Added a new `GlobeLegend` component to display investment categories.


### UI and Structural Changes:
`ElectricityPriceComponent`.
[[1]](diffhunk://#diff-35616f59ef8162269a093fc1be2bc959256eb55058b725af0b4f57a09e2ea82eR6)
[[2]](diffhunk://#diff-35616f59ef8162269a093fc1be2bc959256eb55058b725af0b4f57a09e2ea82eL24-R25)
*
[`app/landingglobe/popups/LandingGlobeNav.tsx`](diffhunk://#diff-0a6879529c47b87e4938dfcf37f46bdb6e91877d643c30d4d4a9c3f6c9bc80b9L136-R142):
Modified the `MainPageNav` component to update the title and description
for better clarity.